Remove step 2f compat shims and consolidate accordion tests (#1346 cleanup)
Post-#1594 cleanup. Removes the two compat shims kept for the soak
period after the accordion + separator/preview module extractions:

  eventHandler.js:
    EventHandler.hasCheckedTemplateGroupToggle = ...
    EventHandler.hasLibraryFileEntries = ...
    EventHandler.highlightParentAccordions = ...
    EventHandler.removeHighlightIfEmpty = ...
    EventHandler.updateAccordionHighlights = ...

  overlayHandler.js:
    OverlayHandler.initializeOverlays = ...
    OverlayHandler.updateHiddenInputs = ...
    OverlayHandler.syncSeparatorPlaceholderFields = ...

Both shims were added as a bridge so classic-script consumers reaching
for window.EventHandler.updateAccordionHighlights (etc.) kept working
during the migration. Grep audit before removal found:

  * ZERO consumers of the 4 EventHandler accordion helpers other than
    the two test files being restructured here.
  * ZERO consumers of any of the 3 OverlayHandler shim methods.
  * ONE consumer of EventHandler.updateAccordionHighlights: an e2e
    smoke test that literally asserted the shim existed. Updated to
    drop the assertion; direct-import coverage of that method already
    exists in the module's own test file.

## Test consolidation

Also relocates the exhaustive accordion-behavior tests from
`tests/js/eventHandler.test.js` (where they lived at the shim's
address, `window.EventHandler.foo(...)`) to
`tests/js/modules/accordionHighlights.test.js` (where they belong,
against direct module imports).

The module test file previously had 14 shallow "does the contract
exist" smoke tests. Those are strict subsets of the 34 exhaustive
tests being moved in from eventHandler.test.js, so they're replaced
rather than merged.

Test file line changes:
  tests/js/eventHandler.test.js:                573 -> 147 lines
    (drops 5 accordion describe blocks; keeps module-registration + toggleLibraryVisibility)
  tests/js/modules/accordionHighlights.test.js: 133 -> 491 lines
    (adds 34 relocated tests + 2 revised surface tests; drops 14 smoke tests)

Test count changes:
  tests/js/eventHandler.test.js:                40 -> 6 tests  (-34)
  tests/js/modules/accordionHighlights.test.js: 14 -> 36 tests (+22)
  Net: -12 tests total (1510 -> 1498)

The -12 delta is entirely the module's shallow smoke tests being
replaced by the deeper suite; no unique coverage is lost. Every
scenario covered by the removed smoke tests is a strict subset of
what the relocated exhaustive tests already cover.

## Files touched

static/local-js/eventHandler.js: -16 net
  - Drops the 5-method compat-shim block (+ its comment)
  - Trims the import from 5 named accordion functions to just the
    one still called internally (updateAccordionHighlights)

static/local-js/overlayHandler.js: -8 net
  - Drops the 3-method compat-shim block (+ its comment)
  - Trims the separator-preview import to drop updateHiddenInputs
    (no longer referenced after the shim removal)

tests/js/eventHandler.test.js: -426 lines (573 -> 147)
  - Removes 5 accordion describe blocks (moved to module test file)
  - Updates "exposes the expected method surface" test to drop the
    5 accordion method names -- they're gone from window.EventHandler
  - Header docstring updated to describe the reduced scope

tests/js/modules/accordionHighlights.test.js: +358 lines (133 -> 491)
  - 6-test module-surface describe block replaced with a 2-test
    version that asserts on the frozen export namespace instead of
    per-name typeof checks (more precise, catches leaks)
  - 8 shallow smoke tests replaced with the 34-test exhaustive suite
    imported from eventHandler.test.js
  - Header docstring updated to reflect the new "canonical test
    location" role

tests/e2e/test_wizard_e2e.py: -3 net
  - Drops the hasHighlights assertion from test_eventhandler_module_loads_via_import
  - Adds a docstring note explaining where the direct-import check lives now

## Verification

- 1498/1498 Vitest pass (1510 -> 1498; -12 from smoke-test dedup)
- 7/7 relevant e2e tests pass:
  * test_eventhandler_module_loads_via_import (the one that used to
    reach for the shim; updated to drop the shim assertion)
  * console-error smoke on all wizard pages
  * library-page interaction tests
  * overlayhandler + eventhandler module-loading tests
- ESLint clean on all 5 touched files

## Result

- window.EventHandler surface: attachLibraryListeners +
  toggleLibraryVisibility only (was ~7 methods before step 2f)
- window.OverlayHandler surface: unchanged from pre-2f minus the
  extracted separator/preview cluster
- Both files now cleanly represent what they actually own; no
  crossover surface remains from the eventHandler <-> overlayHandler
  coupling that motivated the step 2f trilogy.

Step 2f is now completely finished.
